% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_ordpum.R
\name{ordpum}
\alias{ordpum}
\alias{init_y_star_m}
\title{ordpum: A package for sampling from the ordinal probit unfolding model}
\usage{
init_y_star_m(response_v, K)
}
\description{
ordpum: A package for sampling from the ordinal probit unfolding model
}
